remove database_name env var, use stage name as database name move resetdb scripts to test folder so they arent bundled move scripts outside of tifbackendutils so they arent bundled make a new script for resetting the stage db (migration script to come) put all the table definitions in an array so in the future we dont accidentally forget to add one
