+++ .github/workflows/deploy_api.yml
@@ -52,7 +52,7 @@ jobs:
               GITHUB_SHA=${{ github.sha }},\
               ENVIRONMENT=${{ inputs.stage }},\
               DATABASE_HOST=${{ secrets.DATABASE_HOST }},\
-              DATABASE_NAME=${{ secrets.DATABASE_NAME }},\
+              DATABASE_NAME=${{ inputs.stage }},\
               DATABASE_PORT=${{ secrets.DATABASE_PORT }},\
               CA_PEM=${{ secrets.CA_PEM }},\
               DATABASE_USERNAME=$DB_USERNAME,\
