+++ .github/workflows/deploy_geocoder.yml
@@ -47,7 +47,7 @@ jobs:
             --environment "Variables={\
               GITHUB_SHA=${{ github.sha }},\
               DATABASE_HOST=${{ secrets.DATABASE_HOST }},\
-              DATABASE_NAME=${{ secrets.DATABASE_NAME }},\
+              DATABASE_NAME=${{ inputs.stage }},\
               DATABASE_PORT=${{ secrets.DATABASE_PORT }},\
               CA_PEM=${{ secrets.CA_PEM }},\
               DATABASE_USERNAME=$DB_USERNAME,\
