+++ .github/workflows/test_api.yml
@@ -23,7 +23,7 @@ jobs:
         image: mysql:8.0
         env:
           MYSQL_ROOT_PASSWORD: yourpassword
-          MYSQL_DATABASE: tif
+          MYSQL_DATABASE: devTest
         ports:
           - 3306:3306
         options: >-
@@ -37,7 +37,7 @@ jobs:
       DATABASE_PORT: 3306
       DATABASE_USERNAME: root
       DATABASE_PASSWORD: yourpassword
-      DATABASE_NAME: tif
+      DATABASE_NAME: devTest
       AWS_REGION: ${{ secrets.AWS_REGION }}
       AWS_ACCOUNT_ID: ${{ secrets.AWS_ACCOUNT_ID }}
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
@@ -62,10 +62,12 @@ jobs:
       - name: Install dependencies
         run: |
           npm ci
-          cd TiFBackendUtils
-          npm run dbtots
         env:
           CI: true
+
+      - name: Update entities
+        run: |
+          npm run dbtots
       
       - name: Validate openapi specs
         run: |
@@ -99,7 +101,7 @@ jobs:
       CA_PEM: ${{ secrets.CA_PEM }}
       DATABASE_HOST: ${{ secrets.DATABASE_HOST }}
       DATABASE_PORT: ${{ secrets.DATABASE_PORT }}
-      DATABASE_NAME: ${{ secrets.DATABASE_NAME }}
+      DATABASE_NAME: stagingTest
       DATABASE_USERNAME: ${{ secrets.DEV_DATABASE_USERNAME }}
       DATABASE_PASSWORD: ${{ secrets.DEV_DATABASE_PASSWORD }}
       AWS_REGION: ${{ secrets.AWS_REGION }}
@@ -107,7 +109,7 @@ jobs:
       AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
       AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
       ABLY_KEY: ${{ secrets.ABLY_KEY }}
-      AWS_LAMBDA_FUNCTION_NAME: 'lambdaSQLRoute'
+      AWS_LAMBDA_FUNCTION_NAME: lambdaSQLRoute
       COGNITO_CLIENT_APP_ID: ${{ secrets.COGNITO_CLIENT_APP_ID }}
       COGNITO_USER_POOL_ID: ${{ secrets.COGNITO_USER_POOL_ID }}
       SLACK_APP_ID: ${{ secrets.SLACK_APP_ID }}
