+++ .github/workflows/test_geocoder.yml
@@ -16,7 +16,7 @@ jobs:
         image: mysql:8.0
         env:
           MYSQL_ROOT_PASSWORD: yourpassword
-          MYSQL_DATABASE: tif
+          MYSQL_DATABASE: devTest
         ports:
           - 3306:3306
         options: >-
@@ -30,7 +30,7 @@ jobs:
       DATABASE_PORT: 3306
       DATABASE_USERNAME: root
       DATABASE_PASSWORD: yourpassword
-      DATABASE_NAME: tif
+      DATABASE_NAME: devTest
       AWS_REGION: ${{ secrets.AWS_REGION }}
       AWS_ACCOUNT_ID: ${{ secrets.AWS_ACCOUNT_ID }}
       AWS_LAMBDA_FUNCTION_NAME: geocodingPipeline
