+++ APILambda/user/createUser/createUserCheckConstraint.test.ts
@@ -1,6 +1,5 @@
-import { conn } from "TiFBackendUtils"
+import { conn, expectFailsCheckConstraint } from "TiFBackendUtils"
 import { randomUUID } from "crypto"
-import { expectFailsCheckConstraint } from "../../test/database.js"
 import { insertUser } from "./createUserProfile.js"
 
 describe("CheckConstraint tests", () => {
