+++ GeocodingLambda/package.json
@@ -12,7 +12,6 @@
   },
   "scripts": {
     "pr": "cd .. && npm run pr",
-    "dbtots": "node --loader ts-node/esm ../TiFBackendUtils/dbToTs.ts",
     "build": "npx tsc",
     "test": "jest",
     "test:ci": "jest --ci --detectOpenHandles --forceExit --runInBand",
