+++ TiFBackendUtils/index.ts
@@ -1,4 +1,5 @@
 export * from "./AWS/index.js"
+export * from "./Database/index.js"
 export * from "./Retryable/utils.js"
 export * from "./SQLExecutable/index.js"
 export * from "./TiFUserUtils/index.js"
@@ -8,4 +9,3 @@ export * from "./entities.js"
 export * from "./location.js"
 export * from "./result.js"
 export * from "./types/index.js"
-
