+++ TiFBackendUtils/package.json
@@ -18,7 +18,6 @@
   },
   "scripts": {
     "pr": "cd .. && npm run pr",
-    "dbtots": "node --loader ts-node/esm dbToTs.ts",
     "build": "npx tsc",
     "test": "jest",
     "test:ci": "jest --ci --detectOpenHandles --forceExit --runInBand",
